refactor(imghashes): rename images hashes functions
