+++ vhcalc/models/__init__.py
@@ -1,3 +1,9 @@
+from vhcalc.models.imghash_function import ImageHashingFunction
 from vhcalc.models.metadata import MetaData
+from vhcalc.models.url import URL
 
-__all__ = ["MetaData"]
+__all__ = [
+    "MetaData",
+    "URL",
+    "ImageHashingFunction",
+]
