+++ vhcalc/services/__init__.py
@@ -1,7 +1,7 @@
-from .imghashes import a2b_imghash, b2a_imghash, export_imghash_from_media
+from .imghashes import a2b_imghash, b2b_stream_to_imghash, export_imghash_from_media
 
 __all__ = [
     "export_imghash_from_media",
-    "b2a_imghash",
+    "b2b_stream_to_imghash",
     "a2b_imghash",
 ]
